requirements.txt
Flask>=2.2,<3.0
pymysql>=1.0.2
google-generative-ai>=0.2.0
werkzeug>=2.2,<3.0
gunicorn>=20.1.0